+++ src/modules/showComments.js
@@ -1,0 +1,18 @@
+import cardGenerator from './genCommentPopup.js';
+
+const getSurah = async (id) => {
+  const surahURL = `https://api.quran.com/api/v4/chapters/${id}`;
+  const { chapter } = await fetch(surahURL)
+    .then((res) => res.json());
+
+  return chapter;
+};
+
+const commentsPopup = {
+  handleEvent(event) {
+    const parentElem = event.target.parentElement;
+    cardGenerator(getSurah(parentElem.id), parentElem);
+  },
+};
+
+export default commentsPopup;
